Revert "Replaced directory traversal with alphabetical DFS."
This reverts commit 652c12e8bec289dbccc7d92b026a0cf3abe939ef.

Traversing with DFS results in files searched 'in order', however at the cost
about 6% slower runtime (3% slower after normalized to single core). Presumably
due to scandir() needing to call into the kernel, as well as the kernel
dependencies of the free() calls afterwards.

Commit and follow by revert to track this change for future references.